requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.1
django-crispy-forms==2.3
sqlparse==0.5.1
typing_extensions==4.12.2
